pyproject.toml
[project]
name = "llm-uncertainty-quantification-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.8.0",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ipython>=9.7.0",
]
